fix: improve cross-platform correctness and harden Windows path validation
  - Fix PathValidator to block system dirs on any drive letter (not just C:),
    block UNC paths, and allow access to own home dir only (via USERPROFILE)
  - Fix mkdtemp template to use 6 X's as POSIX requires
  - Omit hidden-file dot prefix for temp files on Windows
  - Add VA-API GPU encoding options for Linux in ConverterBridge
  - Replace C-style size_t casts with static_cast in archive write calls
  - Add qtsvg to CI Qt module install

  - Added docs/macos-portability-report.md

+++ src/core/ArchiveConverter.h
@@ -173,7 +173,7 @@ private:
             // Write file contents
             std::ifstream f(filePath.string(), std::ios::binary);
             while ((f.read(buf.data(), buf.size()), f.gcount() > 0))
-                archive_write_data(a, buf.data(), f.gcount());
+                archive_write_data(a, buf.data(), static_cast<size_t>(f.gcount()));
 
             archive_entry_free(entry);
         }
